add  archives and fix error
