--- .gitignore
@@ -1,4 +1,0 @@
-simulador/main.exe
-simulador/promedio.csv
-simulador/puntajes.csv
-simulador/lector.exe
