--- README.md
@@ -1,9 +1,0 @@
-## Computacion Paralela 2020
-Trabajos de computacion paralela y distribuida, Utem 2020
-
-Las carpetas contendran trabajos del Ramo con su explicacion para ser instalada y ejecutada.
-
-#### Clonar todas las carpetas
-```
-git clone https://github.com/BenjaaMartinezG/ComputacionParalela2020.git
-```
