+++ openMP/main.cpp
@@ -5,10 +5,7 @@
 
 using namespace std;
 
-int num_aleatorio(int mayor, int menor)
-{
-  return (rand() % (mayor - (menor + 1)) + menor);
-}
+int num_aleatorio(int mayor, int menor);
 
 int main()
 {
@@ -16,7 +13,7 @@ int main()
   srand((unsigned int)time(0));
   ofstream archivoSalida("puntajes.csv");
 
-  for (int rut = 14916641; rut < 19679045; rut++)
+  for (int rut = 14916641; rut < 19932391; rut++)
   {
     int nem = num_aleatorio(750, 450),
         ranking = num_aleatorio(750, 450),
@@ -37,3 +34,9 @@ int main()
 
     archivoSalida.close();
   }
+}
+
+int num_aleatorio(int mayor, int menor)
+{
+  return (rand() % (mayor - (menor + 1)) + menor);
+}
