+++ openMP/secuencial.cpp
@@ -11,6 +11,7 @@
 
 using namespace std;
 
+//Function
 std::vector<std::string> split(std::string line, char delimiter);
 
 int main(int argc, char **argv)
